// src/enums/LayerName.ts
export enum LayerName {
  PreVoid = "Pre-Void Special",
  Ingot = "Ingot",
  Lagoon = "Lagoon",
  Hilltop = "Hilltop",
  Canyon = "Canyon",
  Divide = "Divide",
  Freezer = "Freezer",
  Forest = "Forest",
  Junction = "Junction",
  Resort = "Resort",
  Tempest = "Tempest",
  Landfill = "Landfill",
  Racing = "Racing",
  Radiance = "Radiance",
  Groove = "Groove",
  Feast = "Feast",
  Forbidden = "Forbidden",
  Core = "Core",
  All = "All of them!"
}